Implement variables
Currently only declaration, initialisation and use are supported, not
assignment. Also fixes a bug with the function resolution code that
would cause user-after-free because a pointer to an element of the
FunctionResolution array was stored in the array. That pointer would
become invalid when the array was resized and thus reallocated. This
was fixed by using a linked list instead.