Implement assignment
For right now this just works for local variables, which are the
only implemented lvalues. But, this should generalise easily to
other cases. check_lvalue.c will probably not be needed once the
type system is in place, as that can be done with const checking
I think. Also fixes a bug with function calling in generate_x64.c
where the argument registers were being initialised with the call
value position rather than the actual argument position.