Fix a few bugs
Firstly, variables with the same name in two different functions
would count as a redeclaration. This was because I wasn't resetting
the num_local_var_resolutions, and also was checking the local var
linked list rather than the array of pointers.

Secondly, I was using the wrong offset to the rbp when access things
on the stack. Apparently I need to also subtract off the size of the
variable, otherwise a segfault occurs. I still haven't fully understood
what's going on, but it seems to be working for now.

Finally I wasn't setting the return value position after a call.